services/orchestrator/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlmodel==0.0.14
psycopg2-binary==2.9.9
httpx==0.26.0
pydantic==2.5.3
python-dotenv==1.0.0
openai==1.10.0
anthropic==0.18.0
